Complete payment integration with invoice generation and email notifications
 Payment Integration:
- Integrated Razorpay payment gateway with test mode fallback
- Added secure payment signature verification
- Support for all payment methods (cards, UPI, net banking, wallets)

 Invoice Generation:
- Professional PDF invoices using ReportLab
- Branded invoices with order details, customer info, itemized products
- Price breakdown with discounts, shipping, and tax

 Email Notifications:
- Automated order confirmation emails with invoice PDF attachment
- Order status update emails
- Beautiful HTML email templates with plain text fallbacks
- Console backend for development testing

 Complete Order Flow:
- Automatic stock reduction on payment
- Coupon usage tracking
- Order history and tracking
- Email sent automatically on payment success

 Testing Tools:
- Management command for testing invoices
- Comprehensive setup guides (PAYMENT_SETUP.md, INTEGRATION_COMPLETE.md)
- Test mode for development without payment keys

 Fixed Issues:
- Fixed product detail template to use 'page' context variable
- Added get_context method to ProductPage for cart form
- Recreated page hierarchy with proper URL structure (/shop/)
- Cart decimal handling for JSON serialization

Files Added:
- shop/invoice.py - PDF generation
- shop/email_utils.py - Email utilities
- shop/management/commands/test_invoice.py - Testing tool
- templates/shop/emails/* - Email templates
- PAYMENT_SETUP.md - Detailed setup guide
- INTEGRATION_COMPLETE.md - Quick start guide
- setup_pages.py - Page creation script

Files Modified:
- shop/models.py - Added email sending to mark_as_paid()
- shop/cart.py - Improved decimal handling
- requirements.txt - Added reportlab and weasyprint
- luvora_project/settings.py - Added SITE_URL config
- .env - Updated with payment and email configs
- templates/shop/product_detail.html - Fixed context variable

+++ shop/cart.py
@@ -91,7 +91,7 @@ class Cart:
         
         for item in cart.values():
             item['price'] = Decimal(item['price'])
-            item['total_price'] = item['price'] * item['quantity']
+            item['total_price'] = float(item['price'] * item['quantity'])  # Convert to float for JSON
             yield item
     
     def __len__(self):
